requirements.txt
iterative_stratification==0.1.7
matplotlib==3.5.1
numpy==1.23.0
pandas==1.4.3
PyYAML==6.0
scikit_learn==1.1.2
scipy==1.8.0
torch==1.12.1
tqdm==4.64.0
ruamel.yaml==0.17.21